add changes smyler suggested
