+++ src/main/java/io/github/terra121/control/fragments/terra/TerraWhereFragment.java
@@ -78,7 +78,7 @@ public class TerraWhereFragment extends CommandFragment {
             return;
         }
         sender.sendMessage(ChatUtil.combine(TextFormatting.GRAY, "Location: ", TextFormatting.BLUE, result[1],
-                TextFormatting.GRAY, ", ", TextFormatting.BLUE, result[0], TextFormatting.GRAY, " Facing: ", TextFormatting.BLUE, GeographicProjection.azimuthToFacing(azimuth), TextFormatting.GRAY, " (", TextFormatting.BLUE, azimuth, TextFormatting.GRAY, ")"));
+                TextFormatting.GRAY, ", ", TextFormatting.BLUE, result[0], TextFormatting.GRAY, " Facing: ", TextFormatting.BLUE, GeographicProjection.azimuthToFacing(azimuth).getRealName(), TextFormatting.GRAY, " (", TextFormatting.BLUE, azimuth, TextFormatting.GRAY, ")"));
     }
 
     @Override
